@@ -14,31 +14,32 @@
     "lint": "eslint ."
   },
   "dependencies": {
-    "bits-ui": "^0.21.12",
     "clsx": "^2.1.1",
-    "lucide-svelte": "^0.446.0",
-    "mode-watcher": "^0.4.0",
-    "svelte-sonner": "^0.3.27",
-    "tailwind-merge": "^2.4.0",
-    "tailwind-variants": "^0.2.1"
+    "tailwind-merge": "^2.5.5",
+    "tailwind-variants": "^0.3.0",
+    "tailwindcss-animate": "^1.0.7"
   },
   "devDependencies": {
-    "@eslint/js": "^9.11.1",
-    "@sveltejs/adapter-auto": "^3.0.0",
-    "@sveltejs/kit": "^2.0.0",
-    "@sveltejs/vite-plugin-svelte": "^3.0.0",
+    "@eslint/js": "^9.16.0",
+    "@sveltejs/adapter-auto": "^3.3.1",
+    "@sveltejs/kit": "^2.9.0",
+    "@sveltejs/vite-plugin-svelte": "^5.0.1",
     "@types/eslint": "^9.6.1",
-    "autoprefixer": "^10.4.19",
-    "eslint": "^9.11.1",
-    "eslint-plugin-svelte": "^2.44.1",
-    "globals": "^15.0.0",
-    "postcss": "^8.4.38",
-    "svelte": "^4.2.7",
-    "svelte-check": "^4.0.4",
-    "tailwindcss": "^3.4.4",
-    "tslib": "^2.4.1",
-    "typescript": "^5.0.0",
-    "typescript-eslint": "^8.8.0",
-    "vite": "^5.0.3"
+    "autoprefixer": "^10.4.20",
+    "bits-ui": "1.0.0-next.67",
+    "eslint": "^9.16.0",
+    "eslint-plugin-svelte": "^2.46.1",
+    "globals": "^15.13.0",
+    "lucide-svelte": "^0.468.0",
+    "mode-watcher": "^0.5.0",
+    "postcss": "^8.4.49",
+    "svelte": "^5.8.1",
+    "svelte-check": "^4.1.1",
+    "svelte-sonner": "^0.3.28",
+    "tailwindcss": "^3.4.16",
+    "tslib": "^2.8.1",
+    "typescript": "^5.7.2",
+    "typescript-eslint": "^8.17.0",
+    "vite": "^6.0.3"
   }
 }
