@@ -1,7 +1,6 @@
 import { Dialog as DialogPrimitive } from "bits-ui";
 
 import Title from "./dialog-title.svelte";
-import Portal from "./dialog-portal.svelte";
 import Footer from "./dialog-footer.svelte";
 import Header from "./dialog-header.svelte";
 import Overlay from "./dialog-overlay.svelte";
@@ -11,6 +10,7 @@ import Description from "./dialog-description.svelte";
 const Root = DialogPrimitive.Root;
 const Trigger = DialogPrimitive.Trigger;
 const Close = DialogPrimitive.Close;
+const Portal = DialogPrimitive.Portal;
 
 export {
 	Root,
