@@ -1,6 +1,6 @@
 import type { Settings } from '$lib/types';
 import { browser } from '$app/environment';
-import { writable } from 'svelte/store';
+import { get, writable } from 'svelte/store';
 
 const defaultSettings: Settings = {
   quickDownloadType: 'off',
@@ -34,6 +34,7 @@ function setSetting<K extends keyof Settings>(key: K, value: Settings[K]) {
     settings[key] = value;
     return settings;
   });
+
   localStorage.setItem('settings', JSON.stringify(settings));
 }
 
