@@ -1,7 +1,3 @@
-<svelte:head>
-    <title>About</title>
-</svelte:head>
-
 <script lang="ts">
   import { onMount } from 'svelte';
   import config from '$lib/config';
@@ -37,49 +33,58 @@
   ];
 </script>
 
+<svelte:head>
+  <title>About</title>
+</svelte:head>
+
 <div class="flex-1 py-12 md:py-24 lg:py-32">
-    <div class="container px-4 md:px-6">
-        <div class="grid gap-12 md:grid-cols-2">
-            <div>
-                <h1 class="text-3xl font-bold sm:text-4xl md:text-5xl">
-                    About
-                </h1>
-                <p class="mt-2 text-muted-foreground md:text-lg">
-                    {config.appName} is a powerful tool that allows you to download videos/photos
-                    from supported platforms.
-                    <br/> Our mission is to provide a free and open-source tool that is easy
-                    to use and accessible to everyone.
-                </p>
+  <div class="container px-4 md:px-6">
+    <div class="grid gap-12 md:grid-cols-2">
+      <div>
+        <h1 class="text-3xl font-bold sm:text-4xl md:text-5xl">About</h1>
+        <p class="mt-2 text-muted-foreground md:text-lg">
+          {config.appName} is a powerful tool that allows you to download videos/photos
+          from supported platforms.
+          <br /> Our mission is to provide a free and open-source tool that is easy
+          to use and accessible to everyone.
+        </p>
+      </div>
+      <div>
+        <h2 class="text-2xl font-bold">Supported Platforms</h2>
+        <p class="mt-1 text-muted-foreground">
+          We are not affiliated with any of the platforms listed below.
+        </p>
+        <div
+          class="grid gap-2 mt-4 grid-cols-{Math.min(
+            3,
+            Math.floor((supportedPlatforms.length - 1) / 5) + 1
+          )}"
+        >
+          {#each supportedPlatforms as platform}
+            <div class="flex items-center">
+              <CheckIcon class="h-5 w-5 text-green-500" />
+              <p class="ml-2 text-muted-foreground">{platform}</p>
             </div>
-            <div>
-                <h2 class="text-2xl font-bold">Supported Platforms</h2>
-                <p class="mt-1 text-muted-foreground">
-                    We are not affiliated with any of the platforms listed below.
-                </p>
-                <div class="grid gap-2 mt-4 grid-cols-{Math.min(3, Math.floor((supportedPlatforms.length - 1) / 5) + 1)}">
-                    {#each supportedPlatforms as platform}
-                        <div class="flex items-center">
-                            <CheckIcon class="h-5 w-5 text-green-500"/>
-                            <p class="ml-2 text-muted-foreground">{platform}</p>
-                        </div>
-                    {/each}
-                </div>
-            </div>
+          {/each}
         </div>
-        <div class="flex flex-col space-y-6 mt-12">
-            {#each bottomButtons as button}
-                <div class="w-full">
-                    <a
-                            href={button.link}
-                            target={button.link.startsWith('#') ? '_self' : '_blank'}
-                            class="text-xl font-bold hover:underline"
-                    >
-                        {button.text}
-                        <ExternalIcon class="h-5 w-5 inline-block ml-0.5 mb-1 ext-primary"/>
-                    </a>
-                    <p class="text-muted-foreground mt-1 mb-2">{button.description}</p>
-                </div>
-            {/each}
-        </div>
+      </div>
     </div>
+    <div class="flex flex-col space-y-6 mt-12">
+      {#each bottomButtons as button}
+        <div class="w-full">
+          <a
+            href={button.link}
+            target={button.link.startsWith('#') ? '_self' : '_blank'}
+            class="text-xl font-bold hover:underline"
+          >
+            {button.text}
+            <ExternalIcon
+              class="h-5 w-5 inline-block ml-0.5 mb-1 ext-primary"
+            />
+          </a>
+          <p class="text-muted-foreground mt-1 mb-2">{button.description}</p>
+        </div>
+      {/each}
+    </div>
+  </div>
 </div>
